+++ src/integration/kotlin/eu/tib/orkg/prototype/paperswithcode/adapters/input/rest/BenchmarkControllerTest.kt
@@ -162,6 +162,46 @@ class BenchmarkControllerTest : RestDocumentationBaseTest() {
     }
 
     @Test
+    fun fetchResearchProblemsForADataset() {
+        val paper = resourceService.create(CreateResourceRequest(null, "Paper", setOf(ClassId("Paper"))))
+
+        val dataset = resourceService.create(CreateResourceRequest(null, "Dataset", setOf(ClassId("Dataset"))))
+
+        val benchmark1 = resourceService.create(CreateResourceRequest(null, "Benchmark 1", setOf(ClassId("Benchmark"))))
+        val benchmark2 = resourceService.create(CreateResourceRequest(null, "Benchmark 2", setOf(ClassId("Benchmark"))))
+
+        val cont1 = resourceService.create(CreateResourceRequest(null, "Contribution 1", setOf(ClassId("Contribution"))))
+        val cont2 = resourceService.create(CreateResourceRequest(null, "Contribution 2", setOf(ClassId("Contribution"))))
+
+        val problem1 = resourceService.create(CreateResourceRequest(null, "Problem 1", setOf(ClassId("Problem"))))
+        val problem2 = resourceService.create(CreateResourceRequest(null, "Problem 2", setOf(ClassId("Problem"))))
+
+        statementService.create(benchmark1.id!!.value, PredicateId("HAS_DATASET"), dataset.id!!.value)
+        statementService.create(benchmark2.id!!.value, PredicateId("HAS_DATASET"), dataset.id!!.value)
+
+        statementService.create(cont1.id!!.value, PredicateId("HAS_BENCHMARK"), benchmark1.id!!.value)
+        statementService.create(cont2.id!!.value, PredicateId("HAS_BENCHMARK"), benchmark2.id!!.value)
+
+        statementService.create(cont1.id!!.value, PredicateId("P32"), problem1.id!!.value)
+        statementService.create(cont1.id!!.value, PredicateId("P32"), problem2.id!!.value)
+        statementService.create(cont2.id!!.value, PredicateId("P32"), problem2.id!!.value)
+
+        statementService.create(paper.id!!.value, PredicateId("P31"), cont1.id!!.value)
+        statementService.create(paper.id!!.value, PredicateId("P31"), cont2.id!!.value)
+
+        mockMvc
+            .perform(getRequestTo("/api/datasets/${dataset.id}/problems"))
+            .andExpect(status().isOk)
+            .andExpect(jsonPath("$", Matchers.hasSize<Int>(2)))
+            .andDo(
+                document(
+                    snippet,
+                    researchProblemListResponseFields()
+                )
+            )
+    }
+
+    @Test
     fun fetchDatasetForResearchProblem() {
         val problem = resourceService.create(CreateResourceRequest(null, "Problem with a dataset", setOf(ClassId("Problem"))))
 
@@ -221,6 +261,9 @@ class BenchmarkControllerTest : RestDocumentationBaseTest() {
     fun fetchDatasetSummary() {
         val dataset = resourceService.create(CreateResourceRequest(null, "some dataset", setOf(ClassId("Dataset"))))
 
+        val problem1 = resourceService.create(CreateResourceRequest(null, "Fancy problem", setOf(ClassId("Problem"))))
+        val problem2 = resourceService.create(CreateResourceRequest(null, "not so fancy problem", setOf(ClassId("Problem"))))
+
         val paper = resourceService.create(CreateResourceRequest(null, "paper", setOf(ClassId("Paper"))))
         val contribution1 = resourceService.create(CreateResourceRequest(null, "Contribution 1", setOf(ClassId("Contribution"))))
         val contribution2 = resourceService.create(CreateResourceRequest(null, "Contribution 2", setOf(ClassId("Contribution"))))
@@ -258,6 +301,9 @@ class BenchmarkControllerTest : RestDocumentationBaseTest() {
             statementService.create(contribution2.id!!.value, PredicateId("HAS_SOURCE_CODE"), it.id!!.value)
         }
 
+        statementService.create(contribution1.id!!.value, PredicateId("P32"), problem1.id!!.value)
+        statementService.create(contribution2.id!!.value, PredicateId("P32"), problem2.id!!.value)
+
         statementService.create(contribution1.id!!.value, PredicateId("HAS_MODEL"), model1.id!!.value)
         statementService.create(contribution2.id!!.value, PredicateId("HAS_MODEL"), model2.id!!.value)
 
@@ -277,9 +323,9 @@ class BenchmarkControllerTest : RestDocumentationBaseTest() {
         statementService.create(evaluationB2E1.id!!.value, PredicateId("HAS_VALUE"), scoreOfM1B2E1.id!!.value)
 
         mockMvc
-            .perform(getRequestTo("/api/datasets/${dataset.id}/summary"))
+            .perform(getRequestTo("/api/datasets/${dataset.id}/problem/${problem1.id}/summary"))
             .andExpect(status().isOk)
-            .andExpect(jsonPath("$", Matchers.hasSize<Int>(3)))
+            .andExpect(jsonPath("$", Matchers.hasSize<Int>(2)))
             .andDo(
                 document(
                     snippet,
@@ -291,6 +337,9 @@ class BenchmarkControllerTest : RestDocumentationBaseTest() {
     @Test
     fun fetchDatasetSummaryWithoutModels() {
         val dataset = resourceService.create(CreateResourceRequest(null, "some dataset", setOf(ClassId("Dataset"))))
+
+        val problem1 = resourceService.create(CreateResourceRequest(null, "Fancy problem", setOf(ClassId("Problem"))))
+        val problem2 = resourceService.create(CreateResourceRequest(null, "not so fancy problem", setOf(ClassId("Problem"))))
 
         val paper = resourceService.create(CreateResourceRequest(null, "paper", setOf(ClassId("Paper"))))
         val contribution1 = resourceService.create(CreateResourceRequest(null, "Contribution 1", setOf(ClassId("Contribution"))))
@@ -315,6 +364,9 @@ class BenchmarkControllerTest : RestDocumentationBaseTest() {
 
         statementService.create(paper.id!!.value, PredicateId("P31"), contribution1.id!!.value)
         statementService.create(paper.id!!.value, PredicateId("P31"), contribution2.id!!.value)
+
+        statementService.create(contribution1.id!!.value, PredicateId("P32"), problem1.id!!.value)
+        statementService.create(contribution2.id!!.value, PredicateId("P32"), problem2.id!!.value)
 
         statementService.create(contribution1.id!!.value, PredicateId("HAS_BENCHMARK"), benchmark1.id!!.value)
         statementService.create(contribution2.id!!.value, PredicateId("HAS_BENCHMARK"), benchmark2.id!!.value)
@@ -342,9 +394,9 @@ class BenchmarkControllerTest : RestDocumentationBaseTest() {
         statementService.create(evaluationB2E1.id!!.value, PredicateId("HAS_VALUE"), scoreOfM1B2E1.id!!.value)
 
         mockMvc
-            .perform(getRequestTo("/api/datasets/${dataset.id}/summary"))
+            .perform(getRequestTo("/api/datasets/${dataset.id}/problem/${problem1.id}/summary"))
             .andExpect(status().isOk)
-            .andExpect(jsonPath("$", Matchers.hasSize<Int>(3)))
+            .andExpect(jsonPath("$", Matchers.hasSize<Int>(2)))
             .andDo(
                 document(
                     snippet,
@@ -368,6 +420,10 @@ class BenchmarkControllerTest : RestDocumentationBaseTest() {
             fieldWithPath("id").description("Research problem ID"),
             fieldWithPath("label").description("Research problem label")
         )
+
+    private fun researchProblemListResponseFields() =
+        responseFields(fieldWithPath("[]").description("A list of research probelms"))
+            .andWithPrefix("[].", researchProblemResponseFields())
 
     private fun benchmarkResponseFields() =
         listOf(